requirements.txt
requests~=2.25.1
beautifulsoup4~=4.10.0
google-api-python-client~=2.42.0
google-auth~=2.6.2
